server/requirements.txt
anyio==3.5.0
APScheduler==3.9.1
asgiref==3.5.0
backports.zoneinfo==0.2.1
black==22.3.0
certifi==2021.10.8
click==8.1.2
decorator==5.1.1
fastapi==0.75.2
h11==0.13.0
idna==3.3
jplephem==2.17
mypy-extensions==0.4.3
numpy==1.22.3
pandas==1.4.2
pathspec==0.9.0
Pillow==9.1.0
platformdirs==2.5.2
py==1.11.0
pydantic==1.9.0
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.4.24
retry==0.9.2
sgp4==2.21
six==1.16.0
skyfield==1.42
sniffio==1.2.0
starlette==0.17.1
tomli==2.0.1
typing-extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
uvicorn==0.17.6
